sweetalert when order success

+++ assets/frontend/js/cart.js
@@ -220,8 +220,28 @@ $(document).ready(function(){
     })
 
     $('.ordernow').click(function(){
-        localStorage.clear();
-        window.location.href="index.php";
+
+        Swal.fire({
+            icon: 'question',
+            title: 'Are you sure?',
+            confirmButtonText: 'Yes',
+            confirmButtonColor: '#0d6efd',
+            showCancelButton: true,
+          }).then((result) => {
+            if (result.isConfirmed) {
+                Swal.fire({
+                    icon: 'success',
+                    title: 'Order Success!<br>Thank you for shopping with us!',
+                    confirmButtonText: 'Got it',
+                    confirmButtonColor: '#0d6efd',
+                  }).then((result) => {
+                    if (result.isConfirmed) {
+                        localStorage.clear();
+                        window.location.href="index.php";
+                    }
+                  })
+            }
+          })
     })
 
     count();
